refactor(db-dynamodb): delete toolbox.ts, relocate types
EntityQueryOptions → utils/entity/types.ts
EntityConstructor, Readonly → utils/entity/Entity.ts (inlined)
DynamoDBTypes → imported directly from EntitySchema.ts
TableDef, ScanOptions → deleted (unused aliases)
