+++ packages/api-websockets-ddb/src/WebsocketsConnectionRegistry.ts
@@ -3,7 +3,7 @@ import { ConnectionRegistry } from "@webiny/api-websockets/exports/api.js";
 import { createEntity } from "./entity.js";
 import type { DynamoDbTableFactory } from "@webiny/db-dynamodb/exports/api/db.js";
 import type { DynamoDbEntityFactory } from "@webiny/db-dynamodb/exports/api/db.js";
-import type { EntityQueryOptions } from "@webiny/db-dynamodb/toolbox.js";
+import type { EntityQueryOptions } from "@webiny/db-dynamodb/exports/api/db.js";
 
 const PK = `WS#CONNECTIONS`;
 const GSI1_PK = "WS#CONNECTIONS#IDENTITY";
