+++ packages/db-dynamodb/src/exports/api/db.ts
@@ -16,6 +16,7 @@ export {
 } from "~/features/DynamoDbEntityFactory/abstractions.js";
 export { DynamoDbBatchFactory } from "~/features/DynamoDbBatchFactory/abstractions.js";
 export type { IEntity } from "~/utils/entity/types.js";
+export type { EntityQueryOptions } from "~/utils/entity/types.js";
 export { ValueFilter } from "~/features/ValueFilter/index.js";
 export { ValueFilterRegistry } from "~/features/ValueFilter/index.js";
 export { FilterUtil } from "~/features/FilterUtil/index.js";
